+++ source/LANDiscovery/LANLobbyManager.cs
@@ -310,6 +310,8 @@ namespace LobbyImprovements.LANDiscovery
         [HarmonyPostfix]
         private static void EnableLeaderboardDisplay(MenuManager __instance, bool enable)
         {
+            __instance.HostSettingsScreen.transform.Find("LIPanel")?.gameObject?.SetActive(!enable);
+
             if (enable && !__instance.requestingLeaderboard && GameNetworkManager.Instance.disableSteam)
             {
                 __instance.requestingLeaderboard = true;