+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+### 1.0.4
+
+- Fixed the lobby list scrollbar being longer than the results
+- Lobby Hosting Changes
+  - Added Password Protected Lobbies
+    - If you have a password set then the lobby will only be joinable by people who have this mod
+  - Added Steam Session Validation
+    - If you have this enabled then the lobby will only be joinable by people who have this mod
+    - If you have this enabled then players without a valid steam session will be disconnected from the lobby
+  - Added the ability to ban players in LAN
+    - If you have this enabled then the lobby will only be joinable by people who have this mod
+  - Added a warning when trying to host a lobby where the name contains a word that is blacklisted in vanilla
+  - Replaced the public, friends-only & invite-only buttons with a dropdown
+- Added ability to set your player name on LAN
+
 ### 1.0.3
 
 - Lobby List Changes